fix(ci): inline MydiaQuality into mix.exs and clean test-env compile
Two CI failures from the first run:

- Docker builds (Build sqlite/postgres, Player E2E) ran 'mix deps.get
  --only prod' and hit '(Code.LoadError) could not load mix_quality.ex'
  because the Dockerfiles copy only mix.exs/mix.lock for dep caching. Move
  the MydiaQuality behaviour-callback predicate inline into mix.exs so it
  is always present when mix.exs is evaluated; delete mix_quality.ex.

- The test-env 'mix compile --warnings-as-errors' gate failed on a
  pre-existing duplicate @doc in test/support/feature_case.ex (fixed in
  its own change).

Also addresses Copilot review on compilers/0: gate :unused to dev/test
(it is a dev/test-only dep). Keeps :unused prepended - appending after
Mix.compilers/0 produces no analysis in this mix_unused version.
